wip move functions and types into modules
