+++ src/command.rs
@@ -1,4 +1,4 @@
-pub(crate) mod list;
 pub(crate) mod build;
 pub(crate) mod clean;
+pub(crate) mod list;
 pub(crate) mod run;
