+++ src/main.rs
@@ -1,250 +1,18 @@
 #![allow(dead_code)]
 
-use std::{collections, env, fmt, fs, io, ops, os, path, process};
+use std::{env, fs, io, os, path};
 
-use clap::{Parser, Subcommand, ValueHint};
-use codespan_reporting::{
-    diagnostic, files,
-    term::{self, termcolor},
+use crate::{
+    cli::{CargoReaperArgs, CargoReaperCommand, Parser},
+    command::{build::build, clean::clean, run::run},
+    error::TomlErrorEmitter,
+    util::{Colorize, PluginManifest, ReaperPluginConfig, validate_plugin},
 };
-use colored::Colorize;
 
-pub(crate) mod util;
-pub(crate) mod error;
+pub(crate) mod cli;
 pub(crate) mod command;
-
-/// Acceptable plugin config toml names for renaming and symlinking REAPER extenion plugins built with Rust.
-pub const CONFIG_FILE_NAMES: &[&str; 2] = &[".reaper.toml", "reaper.toml"];
-
-/// The dynamically linked C library Windows extension
-pub const WINDOWS_PLUGIN_EXT: &str = ".dll";
-/// The dynamically linked C library Linux extension
-pub const LINUX_PLUGIN_EXT: &str = ".so";
-/// The dynamically linked C library MacOS (Darwin) extension
-pub const DARWIN_PLUGIN_EXT: &str = ".dylib";
-
-#[derive(Debug, Parser)]
-#[command(
-    name = "cargo-reaper",
-    version,
-    about = "A Cargo plugin for developing REAPER extension plugins with Rust.",
-    long_about = "`cargo-reaper` is a convenience wrapper around Cargo that adds a post-build hook to streamline REAPER extension development. It automatically renames the compiled plugin to include the required `reaper_` prefix and symlinks it to REAPER’s `UserPlugins` directory.
-
-By default, Cargo prefixes dynamic libraries with `lib`, which REAPER does not recognize. Manually renaming the plugin and keeping the `UserPlugins` directory up-to-date can be tedious -- `cargo-reaper` takes care of all that for you, across all supported platforms."
-)]
-pub struct CargoReaperArgs {
-    #[command(subcommand)]
-    pub(crate) command: CargoReaperCommand,
-}
-#[derive(Debug, Clone, Subcommand)]
-pub enum CargoReaperCommand {
-    /// Create a new REAPER extension plugin from a template at <PATH>.
-    New { path: path::PathBuf },
-
-    /// List available extension plugin(s).
-    List,
-
-    /// Compile REAPER extension plugin(s).
-    Build {
-        /// Do not symlink the plugin(s) to the `UserPlugins` directory.
-        #[arg(long)]
-        no_symlink: bool,
-
-        /// Arguments to forward to the `cargo build` invocation.
-        #[arg(allow_hyphen_values = true, trailing_var_arg = true, num_args = 0.., value_name = "CARGO_BUILD_ARGS")]
-        args: Vec<String>,
-    },
-
-    /// Compile and run REAPER extension plugin(s).
-    Run {
-        /// Override the REAPER executable file path. By default, the REAPER executable found on
-        /// `$PATH` will be used. If the REAPER exectuable can't be found in the current working
-        /// directory, the default global installation path will be used instead.
-        #[arg(long, short = 'e', value_name = "REAPER", value_hint = ValueHint::ExecutablePath)]
-        exec: Option<path::PathBuf>,
-
-        /// Arguments to forward to the `cargo build` invocation.
-        #[arg(allow_hyphen_values = true, trailing_var_arg = true, num_args = 0.., value_name = "CARGO_BUILD_ARGS")]
-        args: Vec<String>,
-    },
-
-    /// Remove plugin(s) from the `UserPlugins` directory that cargo-reaper has generated in the past.
-    Clean {
-        /// Clean plugin(s) by key.
-        #[arg(long = "plugin", short = 'p', value_name = "PLUGIN_KEY")]
-        plugins: Vec<String>,
-
-        /// Display what would be deleted without deleting anything.
-        #[arg(long, short = 'n')]
-        dry_run: bool,
-
-        /// Remove artifacts that cargo-reaper has generated in the past.
-        #[arg(long, short = 'a', default_value = "false")]
-        remove_artifacts: bool,
-    },
-}
-
-/// An identifier that corresponds to some [`codespan_reporting::file::SimpleFile`].
-pub(crate) type FileId = usize;
-
-/// The message to display for some diagnostic error.
-pub(crate) type Message = String;
-
-/// Collection of diagnostics for toml files that is context aware.
-#[derive(Default)]
-pub(crate) struct TomlErrorEmitter<FilePath, FileContents>
-where
-    FilePath: fmt::Display + Clone + Default + Sized,
-    FileContents: AsRef<str> + Clone + Default,
-{
-    /// A collection of file paths and their contents.
-    db: files::SimpleFiles<FilePath, FileContents>,
-
-    /// A collection of diagnostic data containing identifiers corresponding to the db.
-    errors: Vec<diagnostic::Diagnostic<FileId>>,
-}
-impl<FilePath, FileContents> TomlErrorEmitter<FilePath, FileContents>
-where
-    FilePath: fmt::Display + Clone + Default + Sized,
-    FileContents: AsRef<str> + Clone + Default,
-{
-    pub(crate) fn new() -> Self {
-        Default::default()
-    }
-
-    /// Given some diagnostic error info, insert the file and string contents into the db, and
-    /// add an error built from the messages and spans to the list of errors. The file id is
-    /// automatically handled between the diagnostic error and the db.
-    #[allow(clippy::too_many_arguments)]
-    fn insert_err(
-        &mut self,
-        path: FilePath,
-        contents: FileContents,
-        message: impl Into<Message>,
-        primary_span: ops::Range<usize>,
-        primary_msg: Option<impl Into<Message>>,
-        secondary_span: Option<ops::Range<usize>>,
-        secondary_msg: Option<impl Into<Message>>,
-    ) {
-        let error = diagnostic::Diagnostic::error().with_message(message.into());
-        let mut labels: Vec<diagnostic::Label<usize>> = Vec::with_capacity(2);
-        let mut primary_label: diagnostic::Label<usize> = diagnostic::Label::primary(
-            self.db.add(path.clone(), contents.clone()),
-            primary_span.clone(),
-        );
-        if let Some(primary_msg) = primary_msg {
-            primary_label = primary_label.with_message(primary_msg.into());
-        }
-        labels.push(primary_label);
-        if let Some(secondary_msg) = secondary_msg {
-            labels.push(
-                diagnostic::Label::secondary(
-                    self.db.add(path, contents),
-                    secondary_span.unwrap_or(primary_span),
-                )
-                .with_message(secondary_msg.into()),
-            );
-        }
-        self.errors.push(error.with_labels(labels))
-    }
-
-    /// Exit with errors, if any.
-    fn emit(self) -> anyhow::Result<()> {
-        if !self.errors.is_empty() {
-            for error in self.errors.iter().rev() {
-                term::emit(
-                    &mut termcolor::StandardStream::stderr(termcolor::ColorChoice::Auto),
-                    &Default::default(),
-                    &self.db,
-                    error,
-                )?;
-            }
-            process::exit(1);
-        }
-        Ok(())
-    }
-}
-
-/// The parsed contents of a `reaper.toml` config file.
-#[derive(Debug, serde::Deserialize)]
-pub struct ReaperPluginConfig {
-    /// The path to the `reaper.toml` config file.
-    #[serde(skip)]
-    file: path::PathBuf,
-
-    /// The contents of a `reaper.toml` config file as a [`std::str::String`].
-    #[serde(skip)]
-    contents: String,
-
-    /// The contents of a deserialized `reaper.toml` config file.
-    extension_plugins: collections::HashMap<toml::Spanned<String>, toml::Spanned<path::PathBuf>>,
-}
-impl ReaperPluginConfig {
-    /// The path to the `reaper.toml` config file.
-    pub fn file(&self) -> &path::PathBuf {
-        &self.file
-    }
-
-    /// The path to the `reaper.toml` config file.
-    pub fn contents(&self) -> &str {
-        &self.contents
-    }
-
-    /// Locate and deserialize a `reaper.toml` config file.
-    pub fn load(project_root: &path::Path) -> anyhow::Result<Self> {
-        let config_file = CONFIG_FILE_NAMES
-            .iter()
-            .map(|config_file_name| project_root.join(config_file_name))
-            .find(|config_path| config_path.exists())
-            .unwrap(); // We already ensured this path exists
-        let config_contents = fs::read_to_string(&config_file)
-            .map_err(|err| anyhow::anyhow!("failed to read reaper toml file:\n{err:#?}"))?;
-
-        let mut config: Self = toml::from_str(&config_contents).map_err(|err| {
-            anyhow::anyhow!("failed to load plugin config from reaper toml:\n{err:#?}")
-        })?;
-        config.file = config_file;
-        config.contents = config_contents;
-
-        Ok(config)
-    }
-}
-
-/// Represents a REAPER plugin's manifest information.
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct PluginManifest {
-    name: String,
-    version: String,
-    authors: Vec<String>,
-    description: Option<String>,
-}
-impl PluginManifest {
-    fn new(
-        name: String,
-        version: String,
-        authors: Vec<String>,
-        description: Option<String>,
-    ) -> Self {
-        Self {
-            name,
-            version,
-            authors,
-            description,
-        }
-    }
-}
-impl fmt::Display for PluginManifest {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{} v{}", self.name.blue(), self.version)?;
-        if self.description.is_some() {
-            write!(f, " -- {}", self.description.as_ref().unwrap())?;
-        }
-        if !self.authors.is_empty() {
-            write!(f, "\n\nAuthored by: {}", self.authors.join(", "))?;
-        }
-        Ok(())
-    }
-}
+pub(crate) mod error;
+pub(crate) mod util;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -281,8 +49,8 @@ async fn main() -> anyhow::Result<()> {
                 })
         }
         CargoReaperCommand::List => {
-            let mut emitter = TomlErrorEmitter::<String, String>::new();
             let config = ReaperPluginConfig::load(&find_project_root()?)?;
+            let mut emitter = TomlErrorEmitter::<String, String>::new();
             let mut plugins: Vec<String> = Vec::new();
             for (plugin_name, manifest_dir) in config.extension_plugins.iter() {
                 let manifest_file = manifest_dir.get_ref().join("Cargo.toml");
@@ -346,275 +114,6 @@ async fn main() -> anyhow::Result<()> {
             remove_artifacts,
         } => clean(&plugins, dry_run, remove_artifacts),
     }
-}
-
-/// Processes the reaper config toml and the plugin `Cargo.toml` files, collecting diagnostic errors and returning the plugin's manifest.
-pub(crate) fn validate_plugin(
-    emitter: &mut TomlErrorEmitter<String, String>,
-    config_file: &path::Path,
-    config_contents: &str,
-    plugin_name: &toml::Spanned<String>,
-    manifest_file: &path::Path,
-    manifest_file_content: &str,
-) -> anyhow::Result<toml::Spanned<cargo_toml::Manifest>> {
-    let config_file = config_file.to_string_lossy();
-    if !plugin_name.as_ref().starts_with("reaper_") {
-        emitter.insert_err(
-            config_file.to_string(),
-            config_contents.to_string(),
-            "Invalid extension plugin name",
-            plugin_name.span(),
-            Some("extension plugins must be prefixed by `reaper_` to be recognized"),
-            None,
-            Some(format!(
-                "help: consider changing this to `reaper_{}`",
-                plugin_name.as_ref()
-            )),
-        );
-    }
-
-    let manifest = toml::Spanned::new(
-        0..manifest_file_content.len(),
-        cargo_toml::Manifest::from_str(manifest_file_content).map_err(|err| {
-            anyhow::anyhow!(
-                "Failed to parse plugin manifest '{}':\n{err:#?}",
-                manifest_file.display()
-            )
-        })?,
-    );
-
-    let lib = manifest.as_ref().lib.as_ref();
-
-    if lib.is_none() {
-        emitter.insert_err(
-            manifest_file.to_string_lossy().to_string(),
-            manifest_file_content.to_string(),
-            format!(
-                "`{}` does not contain a library target",
-                plugin_name.as_ref()
-            ),
-            manifest.span(),
-            None::<Message>,
-            None,
-            Some("help: add the `[lib]` target attribute"),
-        );
-    }
-    if lib.is_some_and(|lib| lib.name.is_none()) {
-        let lib_index = manifest_file_content.find("[lib]").unwrap();
-        let lib = toml::Spanned::new(lib_index..lib_index + 5, manifest.as_ref().lib.as_ref());
-        emitter.insert_err(
-            manifest_file.to_string_lossy().to_string(),
-            manifest_file_content.to_string(),
-            format!("`{}` library is unnamed", plugin_name.as_ref()),
-            lib.span(),
-            Some("a name is required in order for plugin path resolution during renaming"),
-            None,
-            Some("help: add `name = \"<...>\"`"),
-        );
-    }
-    if lib.is_some_and(|lib| {
-        !lib.crate_type
-            .iter()
-            .any(|crate_type| crate_type == "cdylib")
-    }) {
-        let lib_index = manifest_file_content.find("[lib]").unwrap();
-        let lib = toml::Spanned::new(lib_index..lib_index + 5, manifest.as_ref().lib.as_ref());
-        emitter.insert_err(
-            manifest_file.to_string_lossy().to_string(),
-            manifest_file_content.to_string(),
-            format!("`{}` is not a dynamic library", plugin_name.as_ref()),
-            lib.span(),
-            Some("extension plugins must be dynamic libraries to be recognized"),
-            None,
-            Some("help: add `crate-type = [\"cdylib\"]`"),
-        );
-    }
-    Ok(manifest)
-}
-
-/// Build a REAPER extension plugin.
-pub(crate) fn build(no_symlink: bool, args: Vec<String>) -> anyhow::Result<()> {
-    let project_root = find_project_root()?;
-    let mut emitter = TomlErrorEmitter::<String, String>::new();
-    let config = ReaperPluginConfig::load(&project_root)?;
-
-    match process::Command::new("cargo")
-        .arg("build")
-        .args(&args)
-        .stdin(process::Stdio::inherit())
-        .stdout(process::Stdio::inherit())
-        .stderr(process::Stdio::inherit())
-        .status()
-        .map_err(|err| err.into())
-    {
-        Ok(status) if status.success() => {
-            let profile = args
-                .iter()
-                .find(|arg| *arg == "--release")
-                .map_or("debug", |_| "release");
-
-            for (to_plugin_file_name, plugin_manifest_dir) in config.extension_plugins.iter() {
-                let manifest_file = plugin_manifest_dir.get_ref().join("Cargo.toml");
-                let manifest_file_content = fs::read_to_string(&manifest_file).map_err(|err| {
-                    anyhow::anyhow!(
-                        "Failed to read manifest '{}' for plugin '{}':\n{err:#?}",
-                        manifest_file.display(),
-                        to_plugin_file_name.as_ref()
-                    )
-                })?;
-                let manifest = validate_plugin(
-                    &mut emitter,
-                    config.file(),
-                    config.contents(),
-                    to_plugin_file_name,
-                    &manifest_file,
-                    &manifest_file_content,
-                )?;
-
-                let from_lib_name_with_ext = add_plugin_ext(
-                    &manifest
-                        .into_inner()
-                        .lib
-                        .map(|lib| lib.name.unwrap())
-                        .unwrap(),
-                );
-                let to_lib_name_with_ext = add_plugin_ext(to_plugin_file_name.as_ref());
-                let plugin_path = project_root
-                    .join("target")
-                    .join(profile)
-                    .join(from_plugin_file_name(&from_lib_name_with_ext));
-
-                if plugin_path.exists() {
-                    let plugin_path =
-                        rename_plugin(&project_root, profile, &plugin_path, &to_lib_name_with_ext)?;
-                    if !no_symlink {
-                        symlink_plugin(&plugin_path)?;
-                    } else {
-                        println!(
-                            "{}: plugin was not symlinked ({})",
-                            "warning".yellow().bold(),
-                            plugin_path.display()
-                        );
-                    }
-                }
-            }
-            Ok(())
-        }
-        Ok(status) => {
-            process::exit(status.code().unwrap_or(1));
-        }
-        Err(err) => Err(err),
-    }
-}
-
-/// Remove extension plugins from the `UserPlugins` directory.
-pub(crate) fn clean(
-    plugins: &[String],
-    dry_run: bool,
-    remove_artifacts: bool,
-) -> anyhow::Result<()> {
-    let project_root = find_project_root()?;
-    let config = ReaperPluginConfig::load(&project_root)?;
-    let mut emitter = TomlErrorEmitter::<String, String>::new();
-
-    let plugins: collections::HashMap<String, path::PathBuf> = if !plugins.is_empty() {
-        let mut map = config.extension_plugins;
-        map.retain(|k, _| plugins.contains(k.as_ref()));
-        if map.is_empty() {
-            anyhow::bail!(
-                "The following plugin(s) were not found: {}\n\nTip: run `cargo reaper list` to view the available plugins.",
-                plugins.join(", ")
-            )
-        }
-        map.into_iter()
-            .map(|(key, val)| (key.into_inner(), val.into_inner()))
-            .collect()
-    } else {
-        config
-            .extension_plugins
-            .into_iter()
-            .map(|(key, val)| (key.into_inner(), val.into_inner()))
-            .collect()
-    };
-    let mut removal_failures = 0;
-    for plugin_name in plugins.keys() {
-        println!("    {} {}", "Removing".magenta().bold(), plugin_name);
-        if let Err(err) = remove_plugin_symlink(plugin_name, &add_plugin_ext(plugin_name), dry_run)
-        {
-            removal_failures += 1;
-            eprintln!("{}: {err}", "error (benign)".magenta());
-        }
-    }
-    println!(
-        "     {} {} symlink(s)",
-        if dry_run {
-            "Summary".green().bold()
-        } else {
-            "Removed".green().bold()
-        },
-        plugins.len() - removal_failures
-    );
-
-    if remove_artifacts {
-        let mut package_args: Vec<String> = Vec::with_capacity(plugins.len());
-        for (plugin_name, manifest_dir) in plugins.iter() {
-            let manifest_file = manifest_dir.join("Cargo.toml");
-            let manifest_file_content = fs::read_to_string(&manifest_file).map_err(|err| {
-                anyhow::anyhow!(
-                    "Failed to read manifest '{}' for plugin '{}':\n{err:#?}",
-                    manifest_file.display(),
-                    plugin_name
-                )
-            })?;
-            let mut manifest = toml::Spanned::new(
-                0..manifest_file_content.len(),
-                cargo_toml::Manifest::from_str(&manifest_file_content).map_err(|err| {
-                    anyhow::anyhow!(
-                        "Failed to parse plugin manifest '{}':\n{err:#?}",
-                        manifest_file.display()
-                    )
-                })?,
-            );
-            let _ = manifest
-                .as_mut()
-                .complete_from_path_and_workspace::<cargo_toml::Value>(&manifest_file, None);
-            if let Some(package) = manifest.as_ref().package.as_ref() {
-                package_args.extend(["-p".into(), package.name.clone()]);
-            } else {
-                emitter.insert_err(
-                    manifest_file.to_string_lossy().to_string(),
-                    manifest_file_content,
-                    format!("`{}` is not a package", plugin_name),
-                    manifest.span(),
-                    Some("expected manifest path to a package containing a dynamic library target"),
-                    None,
-                    Some(
-                        "help: is this a workspace? try adding the `[workspace.package]` attribute",
-                    ),
-                );
-            }
-        }
-        emitter.emit()?;
-
-        let mut cargo = process::Command::new("cargo");
-        let mut cargo_clean = cargo
-            .arg("clean")
-            .args(&package_args)
-            .stdin(process::Stdio::inherit())
-            .stdout(process::Stdio::inherit())
-            .stderr(process::Stdio::inherit());
-        if dry_run {
-            cargo_clean = cargo_clean.arg("--dry-run");
-        }
-        cargo_clean.status()?;
-    } else if dry_run {
-        println!(
-            "{}: no files deleted due to --dry-run",
-            "warning".yellow().bold()
-        );
-    }
-
-    Ok(())
 }
 
 /// Rename the resulting extension plugin, returning the new plugin path if it succeeds.
@@ -739,49 +238,6 @@ pub(crate) fn _remove_plugin_symlink(
     )
 }
 
-/// Launch the REAPER binary application. The current working directory takes priority,
-/// but if the binary file is not on `$PATH`, the global default location will be used.
-///
-/// > Note: This function is platform agnostic
-///
-/// # Usage
-///
-/// This is run automatically when running the `cargo reaper run` command.
-fn _run<G>(
-    binary_name: &str,
-    override_binary: Option<path::PathBuf>,
-    run_global_default: G,
-) -> anyhow::Result<()>
-where
-    G: FnOnce() -> io::Result<process::ExitStatus>,
-{
-    override_binary
-        .inspect(|reaper| {
-            println!(
-                "{}: overridng REAPER executable path ({})",
-                "warning".yellow().bold(),
-                reaper.display()
-            )
-        })
-        .or_else(|| which::which(binary_name).ok())
-        .map_or_else(run_global_default, |reaper| {
-            println!(
-                "     {} REAPER executable ({})",
-                "Running".green().bold(),
-                reaper.display(),
-            );
-
-            process::Command::new(reaper)
-                .stdin(process::Stdio::inherit())
-                .stdout(process::Stdio::inherit())
-                .stderr(process::Stdio::inherit())
-                .status()
-        })
-        .map_err(|err| anyhow::anyhow!("While attempting to run REAPER executable: {err:?}"))?;
-
-    Ok(())
-}
-
 #[cfg(target_os = "windows")]
 fn from_plugin_file_name(lib_name: &str) -> String {
     lib_name.to_string()
@@ -834,39 +290,6 @@ pub(crate) fn remove_plugin_symlink(
         dry_run,
     )
 }
-#[cfg(target_os = "windows")]
-fn run(override_binary: Option<path::PathBuf>) -> anyhow::Result<()> {
-    const BINARY_NAME: &str = "reaper";
-    _run(BINARY_NAME, override_binary, || {
-        #[cfg(target_arch = "x86_64")]
-        const GLOBAL_DEFAULT_PATH: &str = r"C:\Program Files\REAPER (x64)\reaper.exe";
-
-        #[cfg(target_arch = "x86")]
-        const GLOBAL_DEFAULT_PATH: &str = r"C:\Program Files (x86)\REAPER\reaper.exe";
-
-        #[cfg(target_arch = "aarch64")]
-        const GLOBAL_DEFAULT_PATH: &str = r"C:\Program Files\REAPER (ARM64)\reaper.exe";
-
-        let reaper = path::PathBuf::from(GLOBAL_DEFAULT_PATH);
-        if reaper.exists() {
-            println!(
-                "     {} global default REAPER executable ({})",
-                "Running".green().bold(),
-                reaper.display(),
-            );
-
-            return process::Command::new(reaper)
-                .stdin(process::Stdio::inherit())
-                .stdout(process::Stdio::inherit())
-                .stderr(process::Stdio::inherit())
-                .status();
-        }
-        Err(io::Error::new(
-            io::ErrorKind::NotFound,
-            "Unable to locate REAPER executable. Is REAPER installed?\n\nTip: Try overriding the default executable path with `--exec <PATH>`.",
-        ))
-    })
-}
 
 #[cfg(target_os = "linux")]
 fn from_plugin_file_name(lib_name: &str) -> String {
@@ -911,27 +334,6 @@ pub(crate) fn remove_plugin_symlink(
             .join("UserPlugins"),
         dry_run,
     )
-}
-#[cfg(target_os = "linux")]
-fn run(override_binary: Option<path::PathBuf>) -> anyhow::Result<()> {
-    const BINARY_NAME: &str = "reaper";
-    _run(BINARY_NAME, override_binary, || {
-        which::which_global(BINARY_NAME)
-            .map_err(|err| io::Error::new(io::ErrorKind::NotFound, err))
-            .and_then(|reaper| {
-                println!(
-                    "     {} global default REAPER executable ({})",
-                    "Running".green().bold(),
-                    reaper.display(),
-                );
-
-                process::Command::new(reaper)
-                    .stdin(process::Stdio::inherit())
-                    .stdout(process::Stdio::inherit())
-                    .stderr(process::Stdio::inherit())
-                    .status()
-            })
-    })
 }
 
 #[cfg(target_os = "macos")]
@@ -981,22 +383,6 @@ pub(crate) fn remove_plugin_symlink(
             .join("UserPlugins"),
         dry_run,
     )
-}
-#[cfg(target_os = "macos")]
-fn run(override_binary: Option<path::PathBuf>) -> anyhow::Result<()> {
-    const BINARY_NAME: &str = "reaper";
-    _run(BINARY_NAME, override_binary, || {
-        const GLOBAL_DEFAULT_ARGS: &[&str; 2] = &["-a", "REAPER"];
-        println!(
-            "     {} global default REAPER executable (/Applications/REAPER.app)",
-            "Running".green().bold(),
-        );
-
-        process::Command::new("open")
-            .args(GLOBAL_DEFAULT_ARGS)
-            .spawn()?
-            .wait()
-    })
 }
 
 fn find_project_root() -> anyhow::Result<path::PathBuf> {
